split lib file into multiple #15
